Update aptly usage with single playbook for multiple distributions
The behavior for the snapshot merge is now handled by a dict which
determines, per distribution, which snapshot has to be merged.

The previous behavior had clashes when two repos where used for
two different releases, and was either duplicated or intersected.

Renaming the snapshots to ignore the distribution introduced all
the changes required for this to work.

On top of that, the aptly playbooks for snapshot creation are now
idempotent.