Move to component instead of prefixes.
The apt pools are separated by their publish prefix. If we use
the tag as prefix, it will keep things well separated, including
the publishing folders, which would lead to more space used.

In this commit, we move to using components instead.
A release tag will therefore appear as a component under the
dist it would use.  We are using still keeping prefixes, but
only to keep the independant (no merge snapshots) and integrated
(merged snapshots) separated.