Add aptly snapshot merge flags
If 2 versions of a the same package for the same architecture in
different repos exist, the order of the merge applies and the merged
snapshot will only get the package (with its possible many versions)
from one repo.

We need this to have versions 1.7 and 2.x for elasticsearch at the same
time in the merged repo.