Add container artifact build tooling
In order to build container artifacts, a chroot
is prepared and the tagfilter strategy plugin
is used to execute the role against the chroot.

The tagfilter plugin facilitates the intelligent
skipping of any tasks with the suffix '-config'
and to also skip the execution of handlers.

With this combination, adding a container to be
built as an artifact it just needs to be added
to the list of roles to process.