Change deployment to consume artifacts
In order to be able to consume artifacts in a
CDC environment, playbooks have been added to
stage the apt, python and container artifacts.

The default deployment (using the deploy.sh
script) implements the staging of only the
python artifacts. The apt and container
artifacts are consumed directly from rpc-repo.

The necessary group_vars to set the
per-container variant to use when executing the
container create role are implemented using
env.d files. Prior to releasing the intent is
to move these settings into the RPC group_vars
using the override folder vars plugin. The
intent of this move will be to ensure that
RPC defaults are kept out of user space as
far as possible.

The rabbitmq_server and galera_server roles
are deliberately executed using the default
container artifact due to initialization issues
(eg: the mariadb root password must be set when
the package is installed) when a pre-installed
container artifact is used. As the packages used
are in the apt artifact repository, this is a
low impact issue.