Use compressed tarballs for git archiving
Using git from rpc-repo to stage the git repos
is failing very often, and the repositories on
rpc-repo are not checked out at the expected SHA
for a release.

This patch switches the git archiving to get the
current checkout, then to archive the checkout
and create a compressed tarball which can be
downloaded using aria2 (which is faster) and
extracted when staging.

This should be faster to stage and more friendly
with corporate firewalls which may not allow
outgoing git traffic through their firewalls.