Adapt AIO apt sources properly
When the apt artifacts are not available (as they have
not yet been built for this release), the AIO bootstrap
must not clobber the updates/backports repository
configuration. This patch ensures that it adapts
accordingly.

A little bit of rpc_release override clean up has also
been removed as it is no longer necessary now that we
are setting the version statically.