Move GROUP_VARS and HOST_VARS_PATH
These variables are now set in the bootstrap-ansible script. This is to prevent
situations where the deploy re-runs bootstrap-ansible.sh as a stand-alone script,
which would overwrite the variables causing out of scope issues.

In addition, some logic has been added so that the variables are set to their
original value before running the OSA bootstrap playbook. This is so there
are no extra settings when the bootstrap playbook runs.