Ensure artifact builds can use common functions
The artifact build process currently fails
because each build script is trying to make
use of the common functions or variables which
are set in the common functions lib, but that
lib is sourced in a subshell.

This patch ensures that the ansible bootstrap
is executed in the same shell so that the env
vars and functions are available to be used.

Also, the git artifact build process now
requires settings from the OSA/RPC group_vars
to complete the artifact build, so this patch
includes ensuring that the group_vars are
sourced appropriately.