Allow py/cnt tests without artifacts
When apt artifacts are not available because
the execution of the script is done as part
of a PR which is a new release, the forced
rewrite of apt configuration must be skipped.

When building the python/container artifacts
before any have been published for the current
release, there is no value for
rpco_mirror_base_url produced by the AIO
bootstrap, so it needs a default to allow the
LXC host playbook to execute and the container
artifact build to progress.