RE-133 Remove previously configured ansible
On the long running apt artifact builder we are
switching Ansible versions for each series, but
leaving behind parts of the installation which
cause the bootstrap to fail.
This is easily seen when doing the ansible
bootstrap using ocata, then trying to bootstrap
using newton. It fails if the role fetch method
is set to use git-clone.

This patch just ensures that we remove
everything related to the ansible execution so
that the new boostrap can do its thing without
worrying about previous installs.