RE-173 Add older ELK apt artifacts
The current ELK deployment makes use of the
older packages. There is a patch in progress [1]
to make use of the newer versions, but it's not
complete yet.

In order to ensure that the current playbooks
can continue to be executed, this patch re-adds
the older versions to the apt artifacting.

This can be reverted once the PR merges.

[1] https://github.com/rcbops/rpc-openstack/pull/2332