RO-3117 Force the correct ownership on aptly files
When synchronising the aptly database from rpc-repo
rsync changes the owner:group to root when it is only
modifying the attributes of the file instead of the
contents - at least I think that's the condition.

After spending several hours trying to figure out
how to do it better in the rsync options, and failing,
this patch opts to use a simpler change which just
forces the owner:group after the sync is completed.

Some line spaces are added into the playbook to improve
readability.