RO-3117 Force the correct owner on artifact files
When synchronising the aptly database to rpc-repo
rsync changes the owner:group to root when it is only
modifying the attributes of the file instead of the
contents - at least I think that's the condition.

After spending several hours trying to figure out
how to do it better in the rsync options, and failing,
this patch opts to use a simpler change which just
forces the owner:group after the sync is completed.

Included are the setting of --no-owner and --nogroup
when doing the pull, and the quoting of the --chown
value. These are important to make it behave correctly
but are unfortunately not enough.