RO-3146 Ensure inventory includes the key file
When the git artifact build wants to upload, the key
file path in the inventory is set to 'False' due to
the REPO_KEYFILE value not being present when the
inventory is prepared near the top of the script.

Given that the inventory is needed to run the initial
playbook, this patch just implements a rebuild of the
inventory if the right env vars are present just before
the upload is actioned.