RO-3284 Set AIO bootstrap options to ensure apt artifacts are used
Currently, when building the python artifacts, the deployment implements
the Ubuntu updates, backports and UCA repositories alongside the rpco apt
artifacts repository. This pollutes the python artifacts with binaries
which are not available in the apt artifact repository which may ultimately
cause service failures when trying to use the python artifacts.

This also affects the container artifact builds as they end up using the
wrong repositories, resulting in builds which differ on every execution.

This patch implements a change to the install script to set the OSA AIO
bootstrap options to prevent it from re-implementing the extra sources.

The functions loading is slightly adjusted so that we can make use of
the OA_OVERRIDES environment variable for the overrides implemented
instead of using yet another override file. The python build script
is then also set to use that variable instead of the file name for
improved consistency.