RO-3284 Either symlink or clone rpc-openstack
In jenkins the folder the script is run from is not
/opt/rpc-openstack in all cases. To cater for this
situation we check whether the folder executed from
has 'rpc-openstack' in the name. If so, we symlink
it to /opt/rpc-openstack. If not, then we think that
rpc-openstack has not been cloned, so we clone it
and replace the submodule with the current directory.