RO-3286 Clean up ansible bootstrapping
Instead of bootstrapping ansible in each build script
just do it in the artifact-setup script. Also clean
up some variable settings which do not need to be
done in bash as they're universal.

We do not need to install python-yaml ourselves any
more because the RPC-O boostrap process does that for
us.