RO-3263 Provide pip extra-links during artifact builds
To build the container artifacts, there is no pip.conf in order
to inform the installation process of the extra links needed to
access the python wheels needed for the installation. This is due
to the use of the tagfilter strategy which skips all -config tags
including the one which places the pip.conf file down.

Unfortunately we cannot use jinja in the strategy to use the config
tag for the default container artifact and not for the rest, so
instead we provide extra pip options here to cover what would be
in pip.conf.