RO-3316 Correct image compression method
When consuming the images prepared with machinectl,
the current method causes machinectl to hang during
the decompression of the image and spikes the CPU
usage to 100%. This affects the RPC-O container
images, but does not happen with upstream images.

In reviewing the difference with how the compression
is done upstream [1], it uses the default compression
rather than our elevated compression and uses a far
simpler set of command-line options.

This patch changes our preparation of the compressed
container images to use the same command line options
for maximum compatibility.

Extra line breaks are added into the playbook in order
to improve readability.

[1] https://github.com/lxc/lxc-ci/blob/9706d80acb6dc2bd3ea519aed82b1b1940953120/bin/build-image#L166