RE-1267 Disable artifacts for new release PR
When doing a PR for a new release, the artifacts do not yet exist
and therefore cannot be used. This patch implements a check in
the scripts which disable the artifacts if they are no available.

The variable set is then re-used where possible to reduce the
number of queries against the artifact server and also to ensure
consistent results.