RE-1267 Correct existing artifact availability checks
Currently the checks run before a python package is installed
to run the script which figures out the rpc_release value. As
such we need to do the checks in the install script after the
installation of the python package, but before the preparation
of the artifact configuration.

This implementation simply disabled their use if they are not
available. As this is the default, it will only take effect if
the artifacts are not available and the artifact build script
intended to use them.